+++ bcbio/structural/shared.py
@@ -15,21 +15,17 @@ from bcbio.bam import callable
 from bcbio.ngsalign import postalign
 from bcbio.pipeline import shared, config_utils
 from bcbio.provenance import do
+from bcbio.variation import population
 
 # ## Case/control
 
 def find_case_control(items):
     """Find case/control items in a population of multiple samples.
     """
-    case_phenotypes = set(["affected"])
-    control_phenotypes = set(["unaffected"])
     cases = []
     controls = []
     for data in items:
-        phenotype = tz.get_in(["metadata", "phenotype"], data)
-        if phenotype in case_phenotypes:
-            cases.append(data)
-        elif phenotype in control_phenotypes:
+        if population.get_affected_status(data) == 1:
             controls.append(data)
         else:
             cases.append(data)